+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.3] - 2025-10-15
+
+### Added
+- Complete recursive search of entire API response to find power data
+- Raw JSON dump of response for analysis
+
 ## [1.1.2] - 2025-10-15
 
 ### Added